done all

+++ app/src/main/java/com/example/aposs_buyer/model/dto/CartDTO.kt
@@ -2,7 +2,7 @@ package com.example.aposs_buyer.model.dto
 
 data class CartDTO(
     var id: Long,
-    var productId: Int,
+    var productId: Long,
     var property: String,
     var quantity: Int,
     var imageUrl: String,
